requirements.txt
streamlit
yfinance
pandas
plotly
groq
python-dotenv